Rename Component.styled.ts => styled.ts for consistency
