+++ src/app/components/Loader/styled.ts
