+++ src/hooks/useDebounce.ts
@@ -1,6 +1,6 @@
 import {useEffect, useState} from 'react';
 
-export function useDebounce<T>(value: T, delay: number = 500): T {
+export const useDebounce = <T>(value: T, delay: number = 500): T => {
   const [debouncedValue, setDebouncedValue] = useState<T>(value);
 
   useEffect(() => {
@@ -14,4 +14,4 @@ export function useDebounce<T>(value: T, delay: number = 500): T {
   }, [value, delay]);
 
   return debouncedValue;
-}
+};
