+++ src/pages/MovieDetailsPage/components/GoBackLink/styled.ts
