+++ src/pages/MovieDetailsPage/styled.ts
