+++ src/pages/MoviesListPage/components/EmptyInitialState/styled.ts
