+++ src/pages/MoviesListPage/components/EmptySearchResult/styled.ts
