+++ src/pages/MoviesListPage/components/MovieCard/MovieCard.tsx
@@ -1,5 +1,5 @@
 import type {Movie} from '../../../../api/movieApi/types';
-import {Card, Content, MetaInfo, Overview, Rating, StyledLink, Title} from './MovieCard.styled';
+import {Card, Content, MetaInfo, Overview, Rating, StyledLink, Title} from './styled';
 import MovieCover from '../../../common/MovieCover/MovieCover';
 import {PathByPageType} from '../../../../routes/constants';
 import {generatePath} from 'react-router-dom';
