+++ src/pages/MoviesListPage/components/MovieCard/styled.ts
