+++ src/pages/MoviesListPage/components/Pagination/styled.ts
