+++ src/pages/MoviesListPage/components/SearchInput/styled.ts
