+++ src/pages/MoviesListPage/styled.ts
