+++ src/pages/common/MovieCover/styled.ts
